requirements.txt
anyio==3.6.1
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
ecdsa==0.17.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
